+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 功能构建
 
+- `webapp-mobile` 子模块实现杂项一级菜单。
+  - 实现外链二级菜单。
+
 - `webapp-mobile` 子模块 hyperscript 页面新增。
   - comvisual/hyperscript/placeholder/index.ts。